improve types

+++ qlat-utils/qlat_utils/include/qlat-utils/array.h
@@ -1,22 +1,22 @@
 #pragma once
 
-#include <cassert>
-
-#include <qlat-utils/show.h>
 #include <qlat-utils/qacc.h>
+#include <qlat-utils/show.h>
+
+#include <cassert>
 
 namespace qlat
 {  //
 
-template <class M, unsigned long N>
+template <class M, size_t N>
 struct API array {
   M v[N];
   //
-  qacc unsigned long size() const { return N; }
+  qacc size_t size() const { return N; }
   //
   qacc void fill(const M& x)
   {
-    for (unsigned long i = 0; i < N; ++i) {
+    for (size_t i = 0; i < N; ++i) {
       v[i] = x;
     }
   }
@@ -24,27 +24,27 @@ struct API array {
   qacc M* data() { return v; }
   qacc const M* data() const { return v; }
   //
-  qacc M& operator[](unsigned long k) { return v[k]; };
-  qacc const M& operator[](unsigned long k) const { return v[k]; };
+  qacc M& operator[](size_t k) { return v[k]; };
+  qacc const M& operator[](size_t k) const { return v[k]; };
 };
 
 template <class M>
 struct API array<M, 0> {
-  qacc unsigned long size() const { return 0; }
+  qacc size_t size() const { return 0; }
   //
   qacc void fill(const M& x) { (void)x; }
   qacc M* data() { return NULL; }
   //
   qacc const M* data() const { return NULL; }
   //
-  qacc M& operator[](unsigned long k)
+  qacc M& operator[](size_t k)
   {
     (void)k;
     assert(false);
     static M x;
     return x;
   };
-  qacc const M& operator[](unsigned long k) const
+  qacc const M& operator[](size_t k) const
   {
     (void)k;
     assert(false);
@@ -53,7 +53,7 @@ struct API array<M, 0> {
   };
 };
 
-template <class M, unsigned long N>
+template <class M, size_t N>
 qacc bool operator<(const array<M, N>& a1, const array<M, N>& a2)
 {
   for (unsigned long i = 0; i < N; ++i) {
@@ -66,7 +66,7 @@ qacc bool operator<(const array<M, N>& a1, const array<M, N>& a2)
   return false;
 }
 
-template <class M, unsigned long N>
+template <class M, size_t N>
 qacc bool operator<=(const array<M, N>& a1, const array<M, N>& a2)
 {
   for (unsigned long i = 0; i < N; ++i) {
